Add arcanist set up to the role
Summary:
This does things a little differently than the old dev-tools role. It still install to .dev-tools
but instead of dropping scripts into /usr/local, it instead makes use of $HOME/.local as the base
script assumes will be used.

For this reason, it also adds a clean up step for /usr/local.

Test Plan: N/A

Reviewers: O5 elhub/devxp, sergei.prodanov

Reviewed By: O5 elhub/devxp, sergei.prodanov

Subscribers: sergei.prodanov

Differential Revision: https://phabricator.elhub.cloud/D6080